Moving Topology to its own module.
